Removed init files

+++ src/test.py
@@ -41,7 +41,7 @@ if __name__ == "__main__":
         reviews = clean_reviews(query_results)
         
         # Construct the prompt with context
-        prompt = generator.build_prompt(query_text, context=reviews, type="short", company_name=company_name)
+        prompt = generator.build_prompt(query_text, context=reviews, type="long", company_name=company_name)
         print("-------------PROMPT VALUE-------------")
         print(f"{prompt}")
